requirements.txt
Flask==0.12.2
mysql-connector-python==8.0.11
numpy==1.14.0